* Extract VisionUtils.computeImageHistogram() and refactored in ReferenceCamera and CameraView. * Split ReferenceCameraWhiteBalanceConfigurationWizard from ReferenceCameraTransformsConfigurationWizard. * Perform Auto-White-Balance actions in machine tasks.
